requirements.txt
alembic==1.9.1
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
orjson==3.8.3
passlib==1.7.4
psycopg2==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
typing-extensions==4.4.0
ujson==5.6.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
zipp==3.11.0
